+++ src/InvalidArgument.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bakame\Aide\Profiler;
+
+use InvalidArgumentException;
+
+final class InvalidArgument extends InvalidArgumentException implements ProfilingException
+{
+}